:sparkles: update cores types

+++ mse_cli/api/types.py
@@ -99,7 +99,7 @@ class Plan(BaseModel):
 
     name: str
     memory: int
-    cores: float
+    cores: int
 
     @staticmethod
     def from_dict(dct: Dict[str, Any]):
